+++ src/wai/json/object/property/_EnumProperty.py
@@ -1,4 +1,4 @@
-from typing import Iterable, Optional
+from typing import Iterable, Optional, Tuple, Set
 
 from ...raw import RawJSONPrimitive
 from ...schema import enum
@@ -14,8 +14,20 @@ class EnumProperty(RawProperty):
                  *,
                  values: Iterable[RawJSONPrimitive] = tuple(),
                  optional: bool = False):
+        # Consume the iterable
+        self._values: Set[RawJSONPrimitive] = set(values)
+
         super().__init__(
             name,
-            schema=enum(*values),
+            schema=enum(*self._values),
             optional=optional
         )
+
+    @property
+    def values(self) -> Set[RawJSONPrimitive]:
+        """
+        Gets the values this property takes.
+
+        :return:    The set of values.
+        """
+        return self._values.copy()