+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Added the ability to create a JSONObject property directly from the class.
 - Properties can now be reused between JSON objects, as long as the attribute
   name stays the same.
+- ConstantProperty and EnumProperty can now report their values.
 
 0.0.1 (2020-02-20)
 -------------------